Moved yuv format guessing to new file and added test.
