+++ YUViewLib/src/video/videoHandlerYUV.h
@@ -71,7 +71,6 @@ public:
 
   // If you know the frame size of the video, the file size (and optionally the bit depth) we can guess
   // the remaining values. The rate value is set if a matching format could be found.
-  // If the sub format is "444" we will assume 4:4:4 input. Otherwise 4:2:0 will be assumed.
   virtual void setFormatFromSizeAndName(const QSize size, int bitDepth, bool packed, int64_t fileSize, const QFileInfo &fileInfo) Q_DECL_OVERRIDE;
 
   // Try to guess and set the format (frameSize/srcPixelFormat) from the raw YUV data.
