+++ YUViewUnitTest/video/video.pro
@@ -3,4 +3,5 @@ TEMPLATE = subdirs
 requires(qtHaveModule(testlib))
 
 SUBDIRS = yuvPixelFormatTest.pro \
-          rgbPixelFormatTest.pro
+          rgbPixelFormatTest.pro \
+          yuvPixelFormatGuessTest.pro
